Refactor to allow reuse of render package
