+++ .gitignore
@@ -16,3 +16,4 @@
 cmd/vaultenv/vaultenv
 cmd/vaultenv/vault-env
 cmd/vaultenv/secret.yml
+cmd/vaultenv/vault-secret.yml
