+++ cmd/vaultenv/Taskfile.yml
@@ -1,0 +1,13 @@
+# https://taskfile.dev
+
+version: '2'
+
+vars:
+  GREETING: Hello, World!
+  APPLICATION: vault-env
+
+tasks:
+  default:
+    cmds:
+      - go build -o {{.APPLICATION}} .
+    silent: true
