+++ cmd/vaultenv/secret.tpl.yml
@@ -1,0 +1,7 @@
+api:
+  key: prod/api.key
+  token: prod/api.token
+cassandra:
+  username: prod/cassandra.username
+  # username: secret/data/prod/cassandra.username
+
